Ist mal wieder Zeit

+++ bundles/edu.kit.kastel.sdq.eclipse.grading.core/src/edu/kit/kastel/sdq/eclipse/grading/core/SystemwideController.java
@@ -6,20 +6,24 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.eclipse.jface.preference.IPreferenceStore;
 
-import edu.kit.kastel.sdq.eclipse.grading.api.IArtemisGUIController;
+import edu.kit.kastel.sdq.eclipse.grading.api.IArtemisController;
 import edu.kit.kastel.sdq.eclipse.grading.api.IAssessmentController;
 import edu.kit.kastel.sdq.eclipse.grading.api.ISystemwideController;
 import edu.kit.kastel.sdq.eclipse.grading.api.PreferenceConstants;
 import edu.kit.kastel.sdq.eclipse.grading.api.alerts.IAlertObservable;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.IProjectFileNamingStrategy;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ICourse;
+import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IExam;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IExercise;
+import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IExerciseGroup;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ISubmission;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ISubmission.Filter;
+import edu.kit.kastel.sdq.eclipse.grading.api.backendstate.Transition;
 import edu.kit.kastel.sdq.eclipse.grading.core.artemis.DefaultProjectFileNamingStrategy;
 import edu.kit.kastel.sdq.eclipse.grading.core.artemis.WorkspaceUtil;
 import edu.kit.kastel.sdq.eclipse.grading.core.config.ConfigDao;
@@ -28,35 +32,46 @@ import edu.kit.kastel.sdq.eclipse.grading.core.config.JsonFileConfigDao;
 public class SystemwideController implements ISystemwideController {
 
 	private final Map<Integer, IAssessmentController> assessmentControllers;
-	private final IArtemisGUIController artemisGUIController;
+	private final IArtemisController artemisGUIController;
 
 	private ConfigDao configDao;
 
 	private Integer courseID;
 	private Integer exerciseID;
 	private Integer submissionID;
-	private String exerciseConfigName;
 
 	private AlertObservable alertObservable;
 
 	private IProjectFileNamingStrategy projectFileNamingStrategy;
 
-	public SystemwideController(final File configFile, final String exerciseConfigName, final String artemisHost, final String username, final String password) {
+	private BackendStateMachine backendStateMachine;
+
+	public SystemwideController(final File configFile, final String artemisHost, final String username, final String password) {
 		this.setConfigFile(configFile);
 		this.assessmentControllers = new HashMap<>();
 		this.alertObservable = new AlertObservable();
 
-		this.exerciseConfigName = exerciseConfigName;
-		this.artemisGUIController = new ArtemisGUIController(this, artemisHost, username, password);
+		this.artemisGUIController = new ArtemisController(this, artemisHost, username, password);
 		this.projectFileNamingStrategy = new DefaultProjectFileNamingStrategy();		//TODO durch das ganze projekt durchreichen! NUR hier instanziieren!
+		this.backendStateMachine = new BackendStateMachine();
 	}
 
 	public SystemwideController(final IPreferenceStore preferenceStore) {
-		this(new File(preferenceStore.getString(PreferenceConstants.P_ABSOLUTE_CONFIG_PATH)),
-				preferenceStore.getString(PreferenceConstants.P_CONFIG_NAME),
-				preferenceStore.getString(PreferenceConstants.P_ARTEMIS_URL),
-				preferenceStore.getString(PreferenceConstants.P_ARTEMIS_USER),
-				preferenceStore.getString(PreferenceConstants.P_ARTEMIS_PASSWORD));
+		this(new File(
+				preferenceStore.getString(PreferenceConstants.ABSOLUTE_CONFIG_PATH)),
+				preferenceStore.getString(PreferenceConstants.ARTEMIS_URL),
+				preferenceStore.getString(PreferenceConstants.ARTEMIS_USER),
+				preferenceStore.getString(PreferenceConstants.ARTEMIS_PASSWORD));
+	}
+
+	private boolean applyTransitionAndNotifyIfNotAllowed(Transition transition) {
+		try {
+			this.backendStateMachine.applyTransition(transition);
+		} catch (NoTransitionException e) {
+			this.alertObservable.error(e.getMessage(), e);
+			return true;
+		}
+		return false;
 	}
 
 	@Override
@@ -65,18 +80,17 @@ public class SystemwideController implements ISystemwideController {
 	}
 
 	@Override
-	public IArtemisGUIController getArtemisGUIController() {
+	public IArtemisController getArtemisGUIController() {
 		return this.artemisGUIController;
 	}
 
-	private IAssessmentController getAssessmentController(int submissionID, String exerciseConfigName, int courseID,
+
+	private IAssessmentController getAssessmentController(int submissionID, int courseID,
 			int exerciseID) {
-		this.assessmentControllers.putIfAbsent(
-				submissionID,
-				new AssessmentController(this, courseID, exerciseID, submissionID, exerciseConfigName));
+		//not equivalent to putIfAbsent!
+		this.assessmentControllers.computeIfAbsent(submissionID, submissionIDParam -> new AssessmentController(this, courseID, exerciseID, submissionID));
 		return this.assessmentControllers.get(submissionID);
 	}
-
 
 	private Collection<ISubmission> getBegunSubmissions(ISubmission.Filter submissionFilter) {
 		if (this.nullCheckMembersAndNotify(true, true, false)) return List.of();
@@ -88,6 +102,15 @@ public class SystemwideController implements ISystemwideController {
 
 	@Override
 	public Collection<String> getBegunSubmissionsProjectNames(ISubmission.Filter submissionFilter) {
+		//sondercase: refresh
+		if (this.courseID == null || this.exerciseID == null) {
+			this.alertObservable.info("You need to choose a"
+				+ (this.courseID == null 							? "course" 		: "")
+				+ (this.courseID == null && this.exerciseID == null ? " and an " 	: "")
+				+ (this.exerciseID == null 							? "exercise" 	: "."));
+			return List.of();
+		}
+
 		return this.getBegunSubmissions(submissionFilter)
 				.stream()
 				.map(submission -> this.projectFileNamingStrategy.getProjectFileInWorkspace(
@@ -108,7 +131,7 @@ public class SystemwideController implements ISystemwideController {
 	@Override
 	public IAssessmentController getCurrentAssessmentController() {
 		if (this.nullCheckMembersAndNotify(true, true, true)) return null;
-		return this.getAssessmentController(this.submissionID, this.exerciseConfigName, this.courseID, this.exerciseID);
+		return this.getAssessmentController(this.submissionID, this.courseID, this.exerciseID);
 	}
 
 	private IExercise getCurrentExercise() {
@@ -118,8 +141,40 @@ public class SystemwideController implements ISystemwideController {
 	}
 
 	@Override
+	public Set<Transition> getCurrentlyPossibleTransitions() {
+		//TODO its best if this is altered based on what correction rounds are enabled
+		boolean secondCorrectionRoundEnabled[] = {false};
+		if (this.exerciseID != null) {
+			this.artemisGUIController.getExercises(this.courseID, true).stream()
+				.filter(exercise -> exercise.getExerciseId() == this.exerciseID)
+				.findAny()
+				.ifPresent(exercise -> secondCorrectionRoundEnabled[0] = exercise.getSecondCorrectionEnabled());
+		}
+		return this.backendStateMachine.getCurrentlyPossibleTransitions().stream()
+				.filter(transition -> !transition.equals(Transition.START_CORRECTION_ROUND_2) || secondCorrectionRoundEnabled[0])
+				.collect(Collectors.toSet());
+	}
+
+	@Override
+	public String getCurrentProjectName() {
+		if (this.nullCheckMembersAndNotify(true, true, true)) return null;
+
+		return this.projectFileNamingStrategy.getProjectFileInWorkspace(
+				WorkspaceUtil.getWorkspaceFile(),
+				this.getCurrentExercise(),
+				this.getCurrentSubmission()).getName();
+	}
+
+	private ISubmission getCurrentSubmission() {
+		if (this.nullCheckMembersAndNotify(true, true, true)) return null;
+		return this.getArtemisGUIController().getSubmissionFromExercise(this.getCurrentExercise(), this.submissionID);
+	}
+
+	@Override
 	public void loadAgain() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.LOAD_AGAIN)) return;
 		if (this.nullCheckMembersAndNotify(true, true, true)) return;
+
 		this.getArtemisGUIController().startAssessment(this.submissionID);
 		this.getArtemisGUIController().downloadExerciseAndSubmission(this.courseID, this.exerciseID, this.submissionID);
 	}
@@ -151,6 +206,8 @@ public class SystemwideController implements ISystemwideController {
 
 	@Override
 	public void onZeroPointsForAssessment() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.ON_ZERO_POINTS_FOR_ASSESSMENT)) return;
+
 		if (this.artemisGUIController.saveAssessment(this.submissionID, true, true)) {
 			this.getCurrentAssessmentController().deleteEclipseProject();
 			this.submissionID = null;
@@ -159,24 +216,24 @@ public class SystemwideController implements ISystemwideController {
 
 	@Override
 	public void reloadAssessment() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.RELOAD_ASSESSMENT)) return;
 		if (this.nullCheckMembersAndNotify(true, true, true)) return;
 
-		this.getCurrentAssessmentController().deleteEclipseProject();
-
-		this.getArtemisGUIController().startAssessment(this.submissionID);
-		this.getArtemisGUIController().downloadExerciseAndSubmission(this.courseID, this.exerciseID, this.submissionID);
-
-		this.getCurrentAssessmentController().resetAndReload();
+		this.getCurrentAssessmentController().resetAndRestartAssessment();
 	}
 
 	@Override
 	public void saveAssessment() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.SAVE_ASSESSMENT)) return;
 		if (this.nullCheckMembersAndNotify(true, true, true)) return;
+
 		this.artemisGUIController.saveAssessment(this.submissionID, false, false);
 	}
 
 	@Override
 	public void setAssessedSubmissionByProjectName(String projectName) {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.SET_ASSESSED_SUBMISSION_BY_PROJECT_NAME)) return;
+
 		boolean[] found = {false};
 		this.getBegunSubmissions(Filter.ALL).forEach(submission -> {
 			String currentProjectName = this.projectFileNamingStrategy.getProjectFileInWorkspace(
@@ -202,6 +259,8 @@ public class SystemwideController implements ISystemwideController {
 
 	@Override
 	public Collection<String> setCourseIdAndGetExerciseShortNames(final String courseShortName) {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.SET_COURSE_ID_AND_GET_EXERCISE_SHORT_NAMES)) return List.of();
+
 		for (ICourse course : this.getArtemisGUIController().getCourses()) {
 			if (course.getShortName().equals(courseShortName)) {
 				this.courseID = course.getCourseId();
@@ -216,47 +275,79 @@ public class SystemwideController implements ISystemwideController {
 
 	@Override
 	public void setExerciseId(final String exerciseShortName) {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.SET_EXERCISE_ID)) return;
+
 		for (ICourse course : this.getArtemisGUIController().getCourses()) {
+			//normal exercises
 			for (IExercise exercise : course.getExercises()) {
 				if (exercise.getShortName().equals(exerciseShortName)) {
 					this.exerciseID = exercise.getExerciseId();
 					return;
 				}
 			}
+			//exam exercises
+			for (IExam exam: course.getExams()) {
+				for (IExerciseGroup exerciseGroup: exam.getExerciseGroups()) {
+					for ( IExercise exercise : exerciseGroup.getExercises()) {
+						if (exercise.getShortName().equals(exerciseShortName)) {
+							this.exerciseID = exercise.getExerciseId();
+							return;
+						}
+					}
+				}
+			}
 		}
+
 		this.alertObservable.error("No Exercise with the given shortName \"" + exerciseShortName + "\" found.", null);
 	}
 
 	@Override
 	public boolean startAssessment() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.START_ASSESSMENT)) return false;
+
 		return this.startAssessment(0);
 	}
 
 	private boolean startAssessment(int correctionRound) {
 		if (this.nullCheckMembersAndNotify(true, true, false)) return false;
+
 		Optional<Integer> optionalSubmissionID = this.getArtemisGUIController().startNextAssessment(this.exerciseID, correctionRound);
 		if (optionalSubmissionID.isEmpty()) {
+			//revert!
+			this.backendStateMachine.revertLatestTransition();
+			this.alertObservable.info("No more submissions available for Correction Round " + (correctionRound + 1) + "!");
 			return false;
 		}
 		this.submissionID = optionalSubmissionID.get();
-		this.getArtemisGUIController().downloadExerciseAndSubmission(this.courseID, this.exerciseID, this.submissionID);
-		return true;
 
+
+		//perform download. Revert state if that fails.
+		if (!this.getArtemisGUIController().downloadExerciseAndSubmission(this.courseID, this.exerciseID, this.submissionID)) {
+			this.backendStateMachine.revertLatestTransition();
+			return false;
+		}
+		return true;
 	}
 
 	@Override
 	public boolean startCorrectionRound1() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.START_CORRECTION_ROUND_1)) return false;
+
 		return this.startAssessment(0);
 	}
 
 	@Override
 	public boolean startCorrectionRound2() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.START_CORRECTION_ROUND_2)) return false;
+
 		return this.startAssessment(1);
 	}
 
 	@Override
 	public void submitAssessment() {
+		if (this.applyTransitionAndNotifyIfNotAllowed(Transition.SUBMIT_ASSESSMENT)) return;
 		if (this.nullCheckMembersAndNotify(true, true, true)) return;
+
 		if (this.artemisGUIController.saveAssessment(this.submissionID, true, false)) {
 			this.getCurrentAssessmentController().deleteEclipseProject();
 			this.assessmentControllers.remove(this.submissionID);
